// app/redux-code/action-types.ts
export enum ActionTypes {
    USER_DATA = "USER_DATA",
    CLEAR_USER = "CLEAR_USER",
    AUTH_TOKEN = "AUTH_TOKEN",
    LOGOUT = "LOGOUT",
    LOGIN = "LOGIN",
    LOGIN_LOADING = "LOGIN_LOADING",
    OTP_VERIFICATION = "OTP_VERIFICATION",
    OTP_LOADING = "OTP_LOADING",
}



export default {};